Fix error when changing case of a selection
This change rolls back some of the refactors made in the previous
commit. While the code could be cleaner, we no longer get error E523
when trying to change the case of a visual selection. The issue seems to
be with the way <expr> was being used, but the specifics are still
unclear.