Move functions to autoload
Thank you very much for a very useful plugin!

**Why** is the change needed?

This is a very neat plugin with a very specific use-case. As such I see
it as a waste of resources to source multiple functions at vim start-up.
I am happy keeping this change in my fork, but I thought you might want
it in the main plugin as well.

**How** is the need addressed?

- Rename functions according to existing naming schema:
  - `s:ActionOpfunc` to `caser#ActionOpfunc`
  - `s:ActionSetup` to `caser#ActionSetup`
  - `s:MapAction` to `caser#MapAction`
  - `s:DoAction` to `caser#DoAction`
- Move them all to the autoload file, trying to reintroduce the same
  folding pattern as previously applied

**What** side effects could this cause?

- Maintainer would have to get used to the change
- The folds and naming might've been different if the maintainer had
  made the change